chore: move SubscriptionData into more appropriate package
