+++ composeApp/src/commonMain/kotlin/com/biondic/rssreader/subscriptions/domain/model/SubscriptionData.kt
@@ -1,9 +1,8 @@
-package com.biondic.rssreader.subscriptions.data.model
+package com.biondic.rssreader.subscriptions.domain.model
 
 import arrow.core.Either
 import com.biondic.rssreader.core.model.DatabaseError
 import com.biondic.rssreader.core.model.NetworkError
-import com.biondic.rssreader.subscriptions.domain.model.Subscription
 
 sealed interface SubscriptionData {
     data class Local(val results: Either<DatabaseError, List<Subscription>>) : SubscriptionData
