+++ composeApp/src/commonMain/kotlin/com/biondic/rssreader/subscriptions/domain/usecase/GetMySubscriptions.kt
@@ -4,7 +4,7 @@ import arrow.core.Either
 import com.biondic.rssreader.core.model.RefreshStrategy
 import com.biondic.rssreader.core.model.RefreshStrategy.SyncWithRemote
 import com.biondic.rssreader.core.model.RepositoryError
-import com.biondic.rssreader.subscriptions.data.model.SubscriptionData.Remote
+import com.biondic.rssreader.subscriptions.domain.model.SubscriptionData.Remote
 import com.biondic.rssreader.subscriptions.domain.model.Subscription
 import com.biondic.rssreader.subscriptions.domain.repo.SubscriptionRepository
 import com.biondic.rssreader.subscriptions.ui.mapper.ExtractSubscriptions
