+++ composeApp/src/commonMain/kotlin/com/biondic/rssreader/subscriptions/ui/mapper/ExtractSubscriptions.kt
@@ -4,9 +4,9 @@ import arrow.core.Either
 import arrow.core.raise.either
 import arrow.core.right
 import com.biondic.rssreader.core.model.RepositoryError
-import com.biondic.rssreader.subscriptions.data.model.SubscriptionData
-import com.biondic.rssreader.subscriptions.data.model.SubscriptionData.Local
-import com.biondic.rssreader.subscriptions.data.model.SubscriptionData.Remote
+import com.biondic.rssreader.subscriptions.domain.model.SubscriptionData
+import com.biondic.rssreader.subscriptions.domain.model.SubscriptionData.Local
+import com.biondic.rssreader.subscriptions.domain.model.SubscriptionData.Remote
 import com.biondic.rssreader.subscriptions.domain.model.Subscription
 
 class ExtractSubscriptions(
